为支持外部管理用户，引入EXTERNALUSER=true的配置表示不在anythingLLM管理用户，通过jwt传入用户id。需要去除外键约束，因为useid是外部的；支持管理员改密码。 引入cross-env支持Windows； 上传的文档备份在服务器； /workspace/new支持传入slug参数； /update-chat支持更新prompt； 开发环境也保存环境设置； 优化vectorSearchResults：先过滤再计数。

+++ collector/utils/files/index.js
@@ -72,8 +72,12 @@ function trashFile(filepath) {
   } catch {
     return;
   }
-
-  fs.rmSync(filepath);
+  if(filepath.indexOf('/hotdir/') === -1)
+    fs.rmSync(filepath);
+  else {
+    const newFn = filepath.replace('/hotdir/', '/docs/')
+    fs.renameSync(filepath, newFn)
+  }
   return;
 }
